Change int to long to avoid possible overflow condition

+++ contrib/platform/src/com/sun/jna/platform/win32/W32Service.java
@@ -242,7 +242,7 @@ public class W32Service implements Closeable {
 				throw new RuntimeException(String.format("Service stop exceeded timeout time of %d ms", timeout));
 			}
 
-			int dwWaitTime = Math.min(sanitizeWaitTime(status.dwWaitHint), (int)msRemainingBeforeTimeout);
+			long dwWaitTime = Math.min(sanitizeWaitTime(status.dwWaitHint), msRemainingBeforeTimeout);
 
 			try {
 				Thread.sleep( dwWaitTime );
